requirements.txt
django<2
django-rest-framework
werkzeug
bcrypt
django-redis
MySQL-python>=1.2.3
requests